refactor(engine): peel merger glob, test-command, and file-scope satellites
Extract conflict patterns/matchGlob, workspace test-command inference, and
file-scope invariant helpers into merger-*.ts modules. Keep re-exports on
merger.ts so existing test and production import paths stay stable.
