+++ packages/engine/src/merger-glob.ts
@@ -1,0 +1,86 @@
+/**
+ * FNXC:CodeOrganization 2026-07-15-12:00:
+ * Conflict path patterns and simple glob matching peeled from merger.ts.
+ * Public path remains merger.ts via re-export for test import stability.
+ */
+/** Conflict type classification for merge conflict resolution */
+export type ConflictType =
+  | "lockfile-ours"
+  | "generated-theirs"
+  | "trivial-whitespace"
+  | "complex";
+
+/** Lock file patterns that should auto-resolve using "ours" (keep current branch's version) */
+export const LOCKFILE_PATTERNS = [
+  "package-lock.json",
+  "pnpm-lock.yaml",
+  "yarn.lock",
+  "Gemfile.lock",
+  "composer.lock",
+  "poetry.lock",
+  "bun.lockb",
+  "go.sum",
+];
+
+/** Generated file patterns that should auto-resolve using "theirs" (keep branch's fresh generation) */
+export const GENERATED_PATTERNS = [
+  "*.gen.ts",
+  "*.gen.js",
+  "*.min.js",
+  "*.min.css",
+  "dist/*",
+  "build/*",
+  "coverage/*",
+  ".next/*",
+  ".nuxt/*",
+  ".output/*",
+  ".cache/*",
+  "out/*",
+  "__generated__/*",
+  "generated/*",
+];
+
+/** Check if a path matches a glob pattern (simple glob support: * and **) */
+export function matchGlob(path: string, pattern: string): boolean {
+  // Handle ** which matches across directory boundaries (must do before single *)
+  if (pattern.includes("**")) {
+    // Convert ** to match any characters including /
+    const regexPattern = pattern
+      .replace(/\./g, "\\.")
+      .replace(/\*\*/g, "<<<DOUBLESTAR>>>")
+      .replace(/\*/g, "[^/]*")
+      .replace(/<<<DOUBLESTAR>>>/g, ".*");
+    const regex = new RegExp(`^${regexPattern}$`);
+    return regex.test(path);
+  }
+  
+  // Handle patterns with single directory wildcards (e.g., "src/*.ts")
+  const lastSlash = pattern.lastIndexOf("/");
+  if (lastSlash !== -1) {
+    const patternDir = pattern.slice(0, lastSlash);
+    const patternFile = pattern.slice(lastSlash + 1);
+    const pathDir = path.lastIndexOf("/") !== -1 ? path.slice(0, path.lastIndexOf("/")) : "";
+    const pathFile = path.lastIndexOf("/") !== -1 ? path.slice(path.lastIndexOf("/")) : path;
+    
+    // Check if directories match
+    if (patternDir.includes("*")) {
+      const dirRegex = new RegExp(`^${patternDir.replace(/\./g, "\\.").replace(/\*/g, "[^/]*")}$`);
+      if (!dirRegex.test(pathDir)) return false;
+    } else if (!pathDir.endsWith(patternDir) && patternDir !== pathDir) {
+      return false;
+    }
+    
+    // Match filename pattern
+    return matchGlob(pathFile, patternFile);
+  }
+  
+  // Simple pattern without directory - match against filename only or full path
+  const fileName = path.lastIndexOf("/") !== -1 ? path.slice(path.lastIndexOf("/") + 1) : path;
+  
+  // Convert glob to regex
+  const regexPattern = pattern
+    .replace(/\./g, "\\.")
+    .replace(/\*/g, "[^/]*");
+  const regex = new RegExp(`^${regexPattern}$`);
+  return regex.test(fileName) || regex.test(path);
+}
